requirements.txt
openai
python-dotenv
websockets
colorama
